pyproject.toml
[project]
name = "agent-data-curation"
version = "0.1.0"
description = "Agentic data curation toolkit for AI-generated image datasets"
requires-python = ">=3.12"
dependencies = [
    "pillow",
    "requests",
    "imagehash",
]

[project.optional-dependencies]
scraping = [
    "scrapling",
    "beautifulsoup4",
    "selenium",
]
playwright = [
    "playwright",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
